+++ source/glbackend/glbackend.h
@@ -363,4 +363,4 @@ void renderBeginScene();
 void renderFinishScene();
 void DrawRateStuff();
 
-void hud_drawsprite(double sx, double sy, int z, int a, int picnum, int dashade, int dapalnum, int dastat);
+void hud_drawsprite(double sx, double sy, int z, double a, int picnum, int dashade, int dapalnum, int dastat);
